+++ ChaoticCorruptions/CHANGELOG.md
@@ -1,3 +1,19 @@
+# 0.5.0
+
+Initial Pre-release, added singularity support
+
+# 0.4.0
+
+Allowed options to craft corrupted cards
+
+# 0.3.0
+
+Allowed options to randomize starting decks
+
+# 0.2.0
+
+Allowed options to corrupt starting decks
+
 # 0.1.0
 
-Initial pre-release. Makes all card rewards corrupted
+Initial testing release. Makes all card rewards corrupted